[IMP] move control buttons to orderlines
