--- addons/point_of_sale/static/src/xml/Screens/ProductScreen/ControlButtons/ProductInfoButton.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-
-    <t t-name="ProductInfoButton" owl="1">
-        <div class="control-button">
-            <i class="fa fa-info-circle" role="img" aria-label="Info" title="Info"/>
-            Info
-        </div>
-    </t>
-
-</templates>
