--- addons/pos_restaurant/static/src/xml/Screens/ProductScreen/ControlButtons/OrderlineNoteButton.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-
-    <t t-name="OrderlineNoteButton" owl="1">
-        <div class="control-button">
-            <i class="fa fa-tag" />
-            <span> </span>
-            <span>Internal Note</span>
-        </div>
-    </t>
-
-</templates>
